refactor UnitList
