+++ apps/dashboard/components/PropertyDetails/MultiFamilyPropertyDetails/UnitList/index.ts
@@ -1,0 +1,1 @@
+export {UnitList} from './UnitList';
